@@ -3,64 +3,51 @@
   "description": "Post from voice for PeerCast to Gitter .",
   "version": "1.0.0",
   "author": "progre",
-  "bugs": "https://github.com/progre/peca-voice/issues",
+  "bugs": {
+    "url": "https://github.com/progre/peca-voice/issues"
+  },
+  "dependencies": {
+    "@types/node": "^8.0.28",
+    "@types/request": "^2.0.8",
+    "jaco": "^2.1.0-beta",
+    "log4js": "^2.3.12",
+    "node-gitter": "^2.1.0",
+    "request": "^2.83.0",
+    "ts-node": "^3.3.0",
+    "tslint": "^5.8.0",
+    "tslint-language-service": "^0.9.6",
+    "twitter": "^1.2.5",
+    "typescript": "^2.6.1"
+  },
   "devDependencies": {
-    "babel-eslint": "^5.0.0",
-    "babel-preset-es2015": "^6.5.0",
-    "babel-preset-modern-node": "^1.0.0",
-    "babelify": "^7.2.0",
-    "browserify": "^13.0.0",
-    "del": "^2.2.0",
-    "gulp": "github:gulpjs/gulp#4.0",
-    "gulp-babel": "^6.1.2",
-    "gulp-concat": "^2.6.0",
-    "gulp-eslint": "^2.0.0",
-    "gulp-espower": "^1.0.2",
-    "gulp-if": "^2.0.0",
-    "gulp-jade": "^1.1.0",
-    "gulp-mocha": "^2.2.0",
-    "gulp-plumber": "^1.1.0",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-stylus": "^2.3.0",
-    "gulp-tslint": "^4.3.2",
-    "gulp-tslint-stylish": "^1.1.1",
-    "gulp-typescript": "^2.11.0",
-    "gulp-uglify": "^1.5.2",
-    "native-promisify": "^1.2.0",
-    "path": "^0.12.7",
-    "power-assert": "^1.2.0",
-    "source-map-support": "^0.4.0",
-    "tsify": "^0.13.2",
-    "tslint": "^3.3.0",
-    "typescript": "^1.7.5",
-    "vinyl-buffer": "^1.0.0",
-    "vinyl-source-stream": "^1.1.0"
+    "@commitlint/cli": "^5.1.1",
+    "@commitlint/config-angular": "^5.1.1",
+    "@progre/tslint-config": "~5.3.0",
+    "husky": "^0.14.3"
+  },
+  "engines": {
+    "node": "8.x.x"
   },
   "files": [
-    "lib"
+    "LICENSE",
+    "src",
+    "tsconfig.json",
+    "tslint.json"
   ],
   "homepage": "https://github.com/progre/peca-voice/",
   "keywords": [
     "template"
   ],
   "license": "MIT",
-  "main": "lib/serve.js",
+  "main": "entry.js",
   "private": true,
-  "repository": "https://github.com/progre/peca-voice.git",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/progre/peca-voice.git"
+  },
   "scripts": {
-    "start": "node .",
-    "now-build": "",
-    "build": "gulp release-build",
-    "test": "gulp test:test"
-  },
-  "dependencies": {
-    "jaco": "^1.2.1",
-    "log4js": "^0.6.31",
-    "node-gitter": "^1.2.9",
-    "request": "^2.69.0",
-    "twitter": "^1.2.5"
-  },
-  "engines": {
-    "node": "4.x.x"
+    "commitmsg": "commitlint -e $GIT_PARAMS",
+    "dev": "ts-node --project tsconfig-dev.json src/serve.ts",
+    "start": "ts-node src/serve.ts"
   }
 }
